update: switch obj loading to library

+++ Engine/engine/Mesh2D.cpp
@@ -7,7 +7,7 @@ void Mesh2D::Draw(VkCommandBuffer commandBuffer) const
 	m_VertexBuffer.Draw(commandBuffer);
 }
 
-void Mesh2D::Load(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices)
+void Mesh2D::Load(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices)
 {
 	m_VertexBuffer.SetVertices(vertices);
 	m_VertexBuffer.SetIndices(indices);
