+++ Engine/engine/Mesh2D.h
@@ -19,7 +19,7 @@ namespace mk
 		Mesh2D& operator=(Mesh2D&& other) noexcept	= delete;
 
 		void Draw(VkCommandBuffer commandBuffer) const;
-		void Load(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices);
+		void Load(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices);
 
 	private:
 		VertexBuffer m_VertexBuffer{};
