+++ Engine/engine/Mesh3D.cpp
@@ -33,7 +33,7 @@ void Mesh3D::Update()
 	m_UniformBuffer.Update(m_WorldTransform);
 }
 
-void Mesh3D::Load(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices)
+void Mesh3D::Load(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices)
 {
 	m_VertexBuffer.SetVertices(vertices);
 	m_VertexBuffer.SetIndices(indices);
@@ -42,7 +42,7 @@ void Mesh3D::Load(const std::vector<Vertex>& vertices, const std::vector<uint16_
 void Mesh3D::Load(const std::string& objFile)
 {
 	std::vector<Vertex> vertices{};
-	std::vector<uint16_t> indices{};
+	std::vector<uint32_t> indices{};
 	if (ParseOBJ<Vertex3D>(objFile, vertices, indices))
 		Load(vertices, indices);
 }
